+++ main/CMakeLists.txt
@@ -14,7 +14,11 @@ idf_component_register(SRCS
                             "apps/settings/settings_sound.c"
                             "managers/app_manager.c" 
                             "apps/launcher/app_launcher.c" 
-                            "overlay_drawer.c" 
+                            "overlay_drawer.c"
+                            "overlay_drawer_ui.c"
+                            "overlay_drawer_control.c"
+                            "overlay_drawer_events.c"
+                            "overlay_drawer_memory.c"
                             "m5stack-tab5-lvgl.c"
                             "gui.c"
                             "hal.c"
