+++ main/overlay_drawer.h
@@ -7,9 +7,9 @@ void app_drawer_open(void);
 void app_drawer_close(void);
 void app_drawer_toggle(void);
 
-// 内存管理函数
+// 内存管理函数（兼容性接口）
 void app_drawer_cleanup_list(void);
 void app_drawer_check_cleanup(void);
 
 // 注册应用抽屉Overlay
-void register_drawer_overlay(void); 
+void register_drawer_overlay(void);
