+++ main/overlay_drawer_control.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "overlay_drawer_types.h"
+
+// 控制面板创建函数
+void drawer_control_create_panel(drawer_state_t* state);
